Time delay for capturing every next frame
The current image capturing is based on time. This capturing method has been changed as it is now done in sampling steps at constant drone speed.
Await time is provided to run animation with delay for frame callback. This file has to be replaced with (drone.js).

Additional changes to the view.js file. Capture function in view.js needs to be replaced with:
capture() {
      
        this.reset();
        this.forest.personMoveTest();
        
        this.drone.capture().then(() => {
            
            this.stage.reset();
            this.export();
        
       
                
        });
          
    }